requirements.txt
absl-py==1.3.0
asn1crypto==1.5.1
astroid==2.12.13
attrs==22.2.0
cached-property==1.5.2
cffi==1.15.1
cfgv==3.3.1
coverage==7.0.0
cryptography==38.0.4
dill==0.3.6
distlib==0.3.6
exceptiongroup==1.0.4
filelock==3.8.2
grpcio==1.51.1
grpcio-tools==1.51.1
identify==2.5.11
importlab==0.8
iniconfig==1.1.1
isort==5.11.4
Jinja2==3.1.2
lazy-object-proxy==1.8.0
libcst==0.4.9
MarkupSafe==2.1.1
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
networkx==2.8.3
ninja==1.11.1
nodeenv==1.7.0
packaging==22.0
platformdirs==2.6.0
pluggy==1.0.0
pre-commit==2.20.0
promise==2.3
protobuf==4.21.12
pycparser==2.21
pycryptodome==3.16.0
pydot==1.4.2
pylint==2.15.9
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
python-pkcs11==0.7.0
pytype==2022.12.15
PyYAML==6.0
six==1.16.0
tabulate==0.9.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
types-protobuf==4.21.0.2
typing-inspect==0.8.0
typing_extensions==4.4.0
virtualenv==20.17.1
wrapt==1.14.1
yapf==0.32.0
